+++ sql/emaj--2.2.2--next_version.sql
@@ -62,6 +62,38 @@ SELECT emaj._disable_event_triggers();
 --                                          --
 ----------------------------------------------
 
+-- index on emaj_mark used to speedup statistics functions, when many marks have been set
+CREATE INDEX emaj_mark_idx1 ON emaj.emaj_mark (mark_time_id);
+
+DROP TYPE emaj.emaj_log_stat_type CASCADE;
+CREATE TYPE emaj.emaj_log_stat_type AS (
+  stat_group                   TEXT,                       -- group name owning the schema.table
+  stat_schema                  TEXT,                       -- schema name
+  stat_table                   TEXT,                       -- table name
+  stat_first_mark              TEXT,                       -- mark representing the lower bound of the time range
+  stat_first_mark_datetime     TIMESTAMPTZ,                -- clock timestamp of the mark representing the lower bound of the time range
+  stat_last_mark               TEXT,                       -- mark representing the upper bound of the time range
+  stat_last_mark_datetime      TIMESTAMPTZ,                -- clock timestamp of the mark representing the upper bound of the time range
+  stat_rows                    BIGINT                      -- estimated number of update events recorded for this table
+  );
+COMMENT ON TYPE emaj.emaj_log_stat_type IS
+$$Represents the structure of rows returned by the emaj_log_stat_group() function.$$;
+
+DROP TYPE emaj.emaj_detailed_log_stat_type CASCADE;
+CREATE TYPE emaj.emaj_detailed_log_stat_type AS (
+  stat_group                   TEXT,                       -- group name owning the schema.table
+  stat_schema                  TEXT,                       -- schema name
+  stat_table                   TEXT,                       -- table name
+  stat_first_mark              TEXT,                       -- mark representing the lower bound of the time range
+  stat_first_mark_datetime     TIMESTAMPTZ,                -- clock timestamp of the mark representing the lower bound of the time range
+  stat_last_mark               TEXT,                       -- mark representing the upper bound of the time range
+  stat_last_mark_datetime      TIMESTAMPTZ,                -- clock timestamp of the mark representing the upper bound of the time range
+  stat_role                    VARCHAR(32),                -- user having generated update events
+  stat_verb                    VARCHAR(6),                 -- type of SQL statement (INSERT/UPDATE/DELETE)
+  stat_rows                    BIGINT                      -- real number of update events recorded for this table
+  );
+COMMENT ON TYPE emaj.emaj_detailed_log_stat_type IS
+$$Represents the structure of rows returned by the emaj_detailed_log_stat_group() function.$$;
 
 --
 -- add created or recreated tables and sequences to the list of content to save by pg_dump
@@ -89,6 +121,253 @@ SELECT emaj._disable_event_triggers();
 ------------------------------------------------------------------
 -- create new or modified functions                             --
 ------------------------------------------------------------------
+CREATE OR REPLACE FUNCTION emaj.emaj_log_stat_group(v_groupName TEXT, v_firstMark TEXT, v_lastMark TEXT)
+RETURNS SETOF emaj.emaj_log_stat_type LANGUAGE plpgsql AS
+$emaj_log_stat_group$
+-- This function returns statistics on row updates executed between 2 marks or between a mark and the current situation.
+-- It is used to quickly get simple statistics of updates logged between 2 marks (i.e. for one or several processing)
+-- These statistics are computed using the serial id of log tables and holes is sequences recorded into emaj_seq_hole at rollback time
+-- Input: group name, the 2 mark names defining a range
+--   a NULL value or an empty string as first_mark indicates the first recorded mark
+--   a NULL value or an empty string as last_mark indicates the current situation
+--   Use a NULL or an empty string as last_mark to know the number of rows to rollback to reach the mark specified by the first_mark parameter.
+--   The keyword 'EMAJ_LAST_MARK' can be used as first or last mark to specify the last set mark.
+-- Output: set of log rows by table (including tables with 0 rows to rollback)
+  DECLARE
+    v_realFirstMark          TEXT;
+    v_realLastMark           TEXT;
+    v_firstMarkTimeId        BIGINT;
+    v_lastMarkTimeId         BIGINT;
+    v_firstMarkTs            TIMESTAMPTZ;
+    v_lastMarkTs             TIMESTAMPTZ;
+  BEGIN
+-- check that the group is recorded in emaj_group table
+    PERFORM 0 FROM emaj.emaj_group WHERE group_name = v_groupName;
+    IF NOT FOUND THEN
+      RAISE EXCEPTION 'emaj_log_stat_group: The group "%" does not exist.', v_groupName;
+    END IF;
+-- if first mark is NULL or empty, retrieve the name, timestamp and last sequ_hole id of the first recorded mark for the group
+    IF v_firstMark IS NULL OR v_firstMark = '' THEN
+--   if no mark exists for the group (just after emaj_create_group() or emaj_reset_group() functions call), v_realFirstMark remains NULL
+      SELECT mark_name, mark_time_id, time_clock_timestamp INTO v_realFirstMark, v_firstMarkTimeId, v_firstMarkTs
+        FROM emaj.emaj_mark, emaj.emaj_time_stamp
+        WHERE time_id = mark_time_id AND mark_group = v_groupName
+        ORDER BY mark_id LIMIT 1;
+    ELSE
+-- else, check and retrieve the name and the timestamp id of the supplied first mark for the group
+      SELECT emaj._get_mark_name(v_groupName,v_firstMark) INTO v_realFirstMark;
+      IF v_realFirstMark IS NULL THEN
+        RAISE EXCEPTION 'emaj_log_stat_group: The start mark "%" is unknown for the group "%".', v_firstMark, v_groupName;
+      END IF;
+      SELECT mark_time_id, time_clock_timestamp INTO v_firstMarkTimeId, v_firstMarkTs
+        FROM emaj.emaj_mark, emaj.emaj_time_stamp
+        WHERE time_id = mark_time_id AND mark_group = v_groupName AND mark_name = v_realFirstMark;
+    END IF;
+-- if a last mark name is supplied, check and retrieve the name, and the timestamp id of the supplied end mark for the group
+    IF v_lastMark IS NOT NULL AND v_lastMark <> '' THEN
+      SELECT emaj._get_mark_name(v_groupName,v_lastMark) INTO v_realLastMark;
+      IF v_realLastMark IS NULL THEN
+        RAISE EXCEPTION 'emaj_log_stat_group: The end mark "%" is unknown for the group "%".', v_lastMark, v_groupName;
+      END IF;
+      SELECT mark_time_id, time_clock_timestamp INTO v_lastMarkTimeId, v_lastMarkTs
+        FROM emaj.emaj_mark, emaj.emaj_time_stamp
+        WHERE time_id = mark_time_id AND mark_group = v_groupName AND mark_name = v_realLastMark;
+-- if last mark is null or empty, v_realLastMark, v_lastMarkTimeId and v_lastMarkTs remain NULL
+    END IF;
+-- check that the first_mark < end_mark
+    IF v_lastMarkTimeId IS NOT NULL AND v_firstMarkTimeId > v_lastMarkTimeId THEN
+      RAISE EXCEPTION 'emaj_log_stat_group: The start mark "%" (%) has been set after the end mark "%" (%).', v_realFirstMark, v_firstMarkTs, v_realLastMark, v_lastMarkTs;
+    END IF;
+-- for each table of the group, get the number of log rows and return the statistics
+-- shorten the timeframe if the table did not belong to the group on the entire requested time frame
+    RETURN QUERY
+      SELECT rel_group, rel_schema, rel_tblseq,
+             CASE WHEN v_firstMarkTimeId IS NULL THEN NULL
+                  WHEN v_firstMarkTimeId >= lower(rel_time_range) THEN v_realFirstMark
+                  ELSE (SELECT mark_name FROM emaj.emaj_mark
+                          WHERE mark_time_id = lower(rel_time_range) AND mark_group = v_groupName)
+               END AS stat_first_mark,
+             CASE WHEN v_firstMarkTimeId IS NULL THEN NULL
+                  WHEN v_firstMarkTimeId >= lower(rel_time_range) THEN v_firstMarkTs
+                  ELSE (SELECT time_clock_timestamp FROM emaj.emaj_time_stamp
+                          WHERE time_id = lower(rel_time_range))
+               END AS stat_first_mark_datetime,
+             CASE WHEN v_lastMarkTimeId IS NULL AND upper_inf(rel_time_range) THEN NULL
+                  WHEN NOT upper_inf(rel_time_range) AND (v_lastMarkTimeId IS NULL OR upper(rel_time_range) < v_lastMarkTimeId)
+                       THEN (SELECT mark_name FROM emaj.emaj_mark
+                               WHERE mark_time_id = upper(rel_time_range) AND mark_group = v_groupName)
+                  ELSE v_realLastMark
+               END AS stat_last_mark,
+             CASE WHEN v_lastMarkTimeId IS NULL AND upper_inf(rel_time_range) THEN NULL
+                  WHEN NOT upper_inf(rel_time_range) AND (v_lastMarkTimeId IS NULL OR upper(rel_time_range) < v_lastMarkTimeId)
+                       THEN (SELECT time_clock_timestamp FROM emaj.emaj_time_stamp
+                               WHERE time_id = upper(rel_time_range))
+                  ELSE v_lastMarkTs
+               END AS stat_last_mark_datetime,
+             CASE WHEN v_firstMarkTimeId IS NULL THEN 0                                              -- group just created but without any mark
+                  ELSE emaj._log_stat_tbl(emaj_relation,
+                                          CASE WHEN v_firstMarkTimeId >= lower(rel_time_range)
+                                                 THEN v_firstMarkTimeId ELSE lower(rel_time_range) END,
+                                          CASE WHEN NOT upper_inf(rel_time_range) AND (v_lastMarkTimeId IS NULL OR upper(rel_time_range) < v_lastMarkTimeId)
+                                                 THEN upper(rel_time_range) ELSE v_lastMarkTimeId END)
+               END AS nb_rows
+        FROM emaj.emaj_relation
+        WHERE rel_group = v_groupName AND rel_kind = 'r'                                             -- tables belonging to the groups
+          AND (upper_inf(rel_time_range) OR upper(rel_time_range) > v_firstMarkTimeId)               --   at the requested time frame
+          AND (v_lastMarkTimeId IS NULL OR lower(rel_time_range) < v_lastMarkTimeId)
+        ORDER BY rel_schema, rel_tblseq, rel_time_range;
+ END;
+$emaj_log_stat_group$;
+COMMENT ON FUNCTION emaj.emaj_log_stat_group(TEXT,TEXT,TEXT) IS
+$$Returns global statistics about logged events for an E-Maj group between 2 marks.$$;
+
+CREATE OR REPLACE FUNCTION emaj.emaj_detailed_log_stat_group(v_groupName TEXT, v_firstMark TEXT, v_lastMark TEXT)
+RETURNS SETOF emaj.emaj_detailed_log_stat_type LANGUAGE plpgsql AS
+$emaj_detailed_log_stat_group$
+-- This function returns statistics on row updates executed between 2 marks as viewed through the log tables
+-- It provides more information than emaj_log_stat_group but it needs to scan log tables in order to provide these data.
+-- So the response time may be much longer.
+-- Input: group name, the 2 marks names defining a range
+--   a NULL value or an empty string as first_mark indicates the first recorded mark
+--   a NULL value or an empty string as last_mark indicates the current situation
+--   The keyword 'EMAJ_LAST_MARK' can be used as first or last mark to specify the last set mark.
+-- Output: table of updates by user and table
+  DECLARE
+    v_realFirstMark          TEXT;
+    v_realLastMark           TEXT;
+    v_firstMarkTimeId        BIGINT;
+    v_lastMarkTimeId         BIGINT;
+    v_firstMarkTs            TIMESTAMPTZ;
+    v_lastMarkTs             TIMESTAMPTZ;
+    v_firstEmajGid           BIGINT;
+    v_lastEmajGid            BIGINT;
+    v_lowerBoundMark         TEXT;
+    v_lowerBoundMarkTs       TIMESTAMPTZ;
+    v_lowerBoundGid          BIGINT;
+    v_upperBoundMark         TEXT;
+    v_upperBoundMarkTs       TIMESTAMPTZ;
+    v_upperBoundGid          BIGINT;
+    v_stmt                   TEXT;
+    r_tblsq                  RECORD;
+    r_stat                   RECORD;
+  BEGIN
+-- check that the group is recorded in emaj_group table
+    PERFORM 0 FROM emaj.emaj_group WHERE group_name = v_groupName;
+    IF NOT FOUND THEN
+      RAISE EXCEPTION 'emaj_detailed_log_stat_group: The group "%" does not exist.', v_groupName;
+    END IF;
+-- if first mark is NULL or empty, retrieve the name, timestamp and last sequ_hole id of the first recorded mark for the group
+    IF v_firstMark IS NULL OR v_firstMark = '' THEN
+      SELECT mark_name, mark_time_id, time_last_emaj_gid, time_clock_timestamp
+        INTO v_realFirstMark, v_firstMarkTimeId, v_firstEmajGid, v_firstMarkTs
+        FROM emaj.emaj_mark, emaj.emaj_time_stamp
+        WHERE time_id = mark_time_id AND mark_group = v_groupName
+        ORDER BY mark_id LIMIT 1;
+    ELSE
+-- else, check and retrieve the name and the timestamp id of the supplied first mark for the group
+      SELECT emaj._get_mark_name(v_groupName,v_firstMark) INTO v_realFirstMark;
+      IF v_realFirstMark IS NULL THEN
+          RAISE EXCEPTION 'emaj_detailed_log_stat_group: The start mark "%" is unknown for the group "%".', v_firstMark, v_groupName;
+      END IF;
+      SELECT mark_name, mark_time_id, time_last_emaj_gid, time_clock_timestamp
+        INTO v_realFirstMark, v_firstMarkTimeId, v_firstEmajGid, v_firstMarkTs
+        FROM emaj.emaj_mark, emaj.emaj_time_stamp
+        WHERE mark_time_id = time_id AND mark_group = v_groupName AND mark_name = v_realFirstMark;
+    END IF;
+-- if no mark has been found (this is the case just after emaj_create_group() or emaj_reset_group() functions call), just exit
+    IF v_realFirstMark IS NULL THEN
+      RETURN;
+    END IF;
+-- catch the timestamp of the last mark
+    IF v_lastMark IS NOT NULL AND v_lastMark <> '' THEN
+-- else, check and retrieve the global sequence value and the timestamp of the end mark for the group
+      SELECT emaj._get_mark_name(v_groupName,v_lastMark) INTO v_realLastMark;
+      IF v_realLastMark IS NULL THEN
+        RAISE EXCEPTION 'emaj_detailed_log_stat_group: The end mark "%" is unknown for the group "%".', v_lastMark, v_groupName;
+      END IF;
+      SELECT mark_time_id, time_last_emaj_gid, time_clock_timestamp INTO v_lastMarkTimeId, v_lastEmajGid, v_lastMarkTs
+        FROM emaj.emaj_mark, emaj.emaj_time_stamp
+        WHERE mark_time_id = time_id AND mark_group = v_groupName AND mark_name = v_realLastMark;
+    END IF;
+-- check that the first_mark < end_mark (v_realFirstMark is known to be not null)
+    IF v_realLastMark IS NOT NULL AND v_firstMarkTimeId > v_lastMarkTimeId THEN
+      RAISE EXCEPTION 'emaj_detailed_log_stat_group: The start mark "%" (%) has been set after the end mark "%" (%).', v_realFirstMark, v_firstMarkTs, v_realLastMark, v_lastMarkTs;
+    END IF;
+-- for each table currently belonging to the group
+    FOR r_tblsq IN
+        SELECT rel_priority, rel_schema, rel_tblseq, rel_time_range, rel_log_schema, rel_log_table FROM emaj.emaj_relation
+          WHERE rel_group = v_groupName AND rel_kind = 'r'                                             -- tables belonging to the groups
+            AND (upper_inf(rel_time_range) OR upper(rel_time_range) > v_firstMarkTimeId)               --   at the requested time frame
+            AND (v_lastMarkTimeId IS NULL OR lower(rel_time_range) < v_lastMarkTimeId)
+          ORDER BY rel_schema, rel_tblseq, rel_time_range
+        LOOP
+-- count the number of operations per type (INSERT, UPDATE and DELETE) and role
+-- compute the lower bound for this table
+      IF v_firstMarkTimeId >= lower(r_tblsq.rel_time_range) THEN
+-- usual case: the table belonged to the group at statistics start mark
+        v_lowerBoundMark = v_realFirstMark;
+        v_lowerBoundMarkTs = v_firstMarkTs;
+        v_lowerBoundGid = v_firstEmajGid;
+      ELSE
+-- special case: the table has been added to the group after the statistics start mark
+        SELECT mark_name INTO v_lowerBoundMark
+          FROM emaj.emaj_mark
+          WHERE mark_time_id = lower(r-tblsq.rel_time_range) AND mark_group = v_groupName;
+        SELECT time_clock_timestamp, time_last_emaj_gid INTO v_lowerBoundMarkTs, v_lowerBoundGid
+          FROM emaj.emaj_time_stamp
+          WHERE time_id = lower(r_tblsq.rel_time_range);
+      END IF;
+-- compute the upper bound for this table
+      IF v_lastMarkTimeId IS NULL AND upper_inf(r_tblsq.rel_time_range) THEN
+-- no supplied end mark and the table has not been removed from its group => the current situation
+        v_upperBoundMark = NULL;
+        v_upperBoundMarkTs = NULL;
+        v_upperBoundGid = NULL;
+      ELSIF NOT upper_inf(r_tblsq.rel_time_range) AND (v_lastMarkTimeId IS NULL OR upper(r_tblsq.rel_time_range) < v_lastMarkTimeId) THEN
+-- special case: the table has been removed from its group before the statistics end mark
+        SELECT mark_name INTO v_upperBoundMark
+          FROM emaj.emaj_mark
+          WHERE mark_time_id = upper(r_tblsq.rel_time_range) AND mark_group = v_groupName;
+        SELECT time_clock_timestamp, time_last_emaj_gid INTO v_upperBoundMarkTs, v_upperBoundGid
+          FROM emaj.emaj_time_stamp
+          WHERE time_id = upper(r_tblsq.rel_time_range);
+      ELSE
+-- usual case: the table belonged to the group at statistics end mark
+        v_upperBoundMark = v_realLastMark;
+        v_upperBoundMarkTs = v_lastMarkTs;
+        v_upperBoundGid = v_lastEmajGid;
+      END IF;
+-- build the statement
+      v_stmt= 'SELECT ' || quote_literal(v_groupName) || '::TEXT AS stat_group, '
+           || quote_literal(r_tblsq.rel_schema) || '::TEXT AS stat_schema, '
+           || quote_literal(r_tblsq.rel_tblseq) || '::TEXT AS stat_table, '
+           || quote_literal(v_lowerBoundMark) || '::TEXT AS stat_first_mark, '
+           || quote_literal(v_lowerBoundMarkTs) || '::TIMESTAMPTZ AS stat_first_mark_datetime, '
+           || coalesce(quote_literal(v_upperBoundMark),'NULL') || '::TEXT AS stat_last_mark, '
+           || coalesce(quote_literal(v_upperBoundMarkTs),'NULL') || '::TIMESTAMPTZ AS stat_last_mark_datetime, '
+           || ' emaj_user AS stat_user,'
+           || ' CASE emaj_verb WHEN ''INS'' THEN ''INSERT'''
+           ||                ' WHEN ''UPD'' THEN ''UPDATE'''
+           ||                ' WHEN ''DEL'' THEN ''DELETE'''
+           ||                             ' ELSE ''?'' END::VARCHAR(6) AS stat_verb,'
+           || ' count(*) AS stat_rows'
+           || ' FROM ' || quote_ident(r_tblsq.rel_log_schema) || '.' || quote_ident(r_tblsq.rel_log_table)
+           || ' WHERE NOT (emaj_verb = ''UPD'' AND emaj_tuple = ''OLD'')'
+           || ' AND emaj_gid > '|| v_lowerBoundGid
+           || coalesce(' AND emaj_gid <= '|| v_upperBoundGid, '')
+           || ' GROUP BY stat_group, stat_schema, stat_table, stat_user, stat_verb'
+           || ' ORDER BY stat_user, stat_verb';
+-- and execute the statement
+      FOR r_stat IN EXECUTE v_stmt LOOP
+        RETURN NEXT r_stat;
+      END LOOP;
+    END LOOP;
+    RETURN;
+  END;
+$emaj_detailed_log_stat_group$;
+COMMENT ON FUNCTION emaj.emaj_detailed_log_stat_group(TEXT,TEXT,TEXT) IS
+$$Returns detailed statistics about logged events for an E-Maj group between 2 marks.$$;
+
 --<end_functions>                                pattern used by the tool that extracts and insert the functions definition
 ------------------------------------------
 --                                      --
@@ -111,6 +390,8 @@ GRANT SELECT ON ALL TABLES IN SCHEMA emaj TO emaj_viewer;
 GRANT SELECT ON ALL SEQUENCES IN SCHEMA emaj TO emaj_viewer;
 REVOKE SELECT ON TABLE emaj.emaj_param FROM emaj_viewer;
 
+GRANT EXECUTE ON FUNCTION emaj.emaj_log_stat_group(v_groupName TEXT, v_firstMark TEXT, v_lastMark TEXT) TO emaj_viewer;
+GRANT EXECUTE ON FUNCTION emaj.emaj_detailed_log_stat_group(v_groupName TEXT, v_firstMark TEXT, v_lastMark TEXT) TO emaj_viewer;
 
 ------------------------------------
 --                                --